pyproject.toml
[tool.poetry]
name = "kurs"
version = "0.1.0"
description = ""
authors = ["миша <misatoropkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
black = "^24.10.0"
mypy = "^1.14.1"
isort = "^5.13.2"
flake8 = "^7.1.1"
pytest = "^8.3.4"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
